Add files via upload
The poetry generation process involves selecting a starting word for each verse randomly and then iteratively predicting subsequent words based on the n-gram models. While the primary focus is on bigram models due to computational simplicity, the codebase is designed to support experimentation with trigram models and potentially more sophisticated techniques in the future.

This repository serves as a valuable resource for both enthusiasts of Urdu poetry and researchers interested in natural language processing and computational creativity. Join us in exploring the intersection of literature and technology, and unlock the poetic potential of n-gram language modeling in Urdu poetry generation!"